Add files via upload
So I managed accomplish the following in this code; Add A SpinBox to UI to set rotation value, pass the value to the Selected Layout and Rubberband, reload the rubberband with new rotation value on rotation value change, cleaned unused import lines, added comments where I added code.
The two remaining issues are; 1). I can only get it to rotate on the Upper Left corner of the rectangle. When I try to use a 0.5*...width() function on various rectangles/extents the point of rotation becomes inconsistent depending on map scale and position. If I use a center.x()/y() approach to the rotation it seems to rotate based on a geographic location that is nowhere near the center of the rectangel being generated. I thought this would be the best way to share my work for those who might want to try and tackle those issues. I am still working on it as well and will update if I find a solution.
Thanks again for this great tool;
Tyler